+++ module/common/get_extra.sh
@@ -3,7 +3,6 @@
 # This file is the backend of JavaScript
 
 MODPATH=${0%/*}
-ORG_PATH="$PATH"
 SKIPLIST="$MODPATH/tmp/skiplist"
 XPOSED="$MODPATH/tmp/xposed"
 
@@ -137,8 +136,73 @@ set_security_patch() {
 }
 
 get_latest_security_patch() {
-    security_patch=$(download "https://source.android.com/docs/security/bulletin/pixel" | grep -o "<td>[0-9]\{4\}-[0-9]\{2\}-[0-9]\{2\}</td>" | head -n 1 | sed 's/<td>\(.*\)<\/td>/\1/')
-    [ -n "$security_patch" ] && echo "$security_patch" || exit 1
+    security_patch=$(download "https://source.android.com/docs/security/bulletin/pixel" |
+                     sed -n 's/.*<td>\([0-9]\{4\}-[0-9]\{2\}-[0-9]\{2\}\)<\/td>.*/\1/p' |
+                     head -n 1)
+
+    if [ -n "$security_patch" ]; then
+        echo "$security_patch"
+        exit 0
+    elif ! ping -c 1 -W 5 "source.android.com" >/dev/null 2>&1; then
+        echo "Connection failed" >&2
+    fi
+    exit 1
+}
+
+unknown_kb() {
+    # adapted from https://github.com/TMLP-Team/keyboxGenerator/blob/main/keyboxGenerator_v2.0.py
+    ECKEY="eckey.pem"
+    CERT="cert.pem"
+    RSAKEY="rsakey.pem"
+    KEYBOX="keybox.xml"
+
+    # gen ec_key
+    openssl ecparam -name prime256v1 -genkey -noout -out "$ECKEY" || exit 1
+
+    # gen cert
+    openssl req -new -x509 -key "$ECKEY" -out "$CERT" -days 3650 -subj "/CN=Generated" || exit 1
+
+    # gen rsa key
+    openssl genrsa -out "$RSAKEY" 2048 || exit 1
+
+    # convert rsa key to PKCS#1
+    openssl rsa -in "$RSAKEY" -out "$RSAKEY" -traditional || exit 1
+
+    # Generate keybox XML
+    cat << KEYBOX_EOF > "$KEYBOX"
+<?xml version="1.0"?>
+    <AndroidAttestation>
+        <NumberOfKeyboxes>1</NumberOfKeyboxes>
+        <Keybox DeviceID="sw">
+            <Key algorithm="ecdsa">
+                <PrivateKey format="pem">
+$(sed 's/^/                    /' "$ECKEY")
+                </PrivateKey>
+                <CertificateChain>
+                    <NumberOfCertificates>1</NumberOfCertificates>
+                        <Certificate format="pem">
+$(sed 's/^/                        /' "$CERT")
+                        </Certificate>
+                </CertificateChain>
+            </Key>
+            <Key algorithm="rsa">
+                <PrivateKey format="pem">
+$(sed 's/^/                    /' "$RSAKEY")
+                </PrivateKey>
+            </Key>
+        </Keybox>
+</AndroidAttestation>
+KEYBOX_EOF
+
+    # cleanup
+    rm -f $ECKEY $CERT $RSAKEY
+
+    if [ -f $KEYBOX ]; then
+        mv /data/adb/tricky_store/keybox.xml /data/adb/tricky_store/keybox.xml.bak
+        mv "$KEYBOX" /data/adb/tricky_store/keybox.xml
+    else
+        exit 1
+    fi
 }
 
 case "$1" in
@@ -184,6 +248,10 @@ case "$1" in
     ;;
 --get-security-patch)
     get_latest_security_patch
+    exit
+    ;;
+--unknown-kb)
+    unknown_kb
     exit
     ;;
 esac